% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_program.R
\name{dynamic_program}
\alias{dynamic_program}
\title{Dynamic_program}
\usage{
dynamic_program(x, w)
}
\arguments{
\item{x}{A DataFrame with two columns (weight(w) and value(v))}

\item{w}{Sack maximum weight capacity}
}
\value{
rreturns item with maximized value within knapsack weight limit
}
\description{
Dynamic_program
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
