% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_approach.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{greedy_knapsack}
\usage{
greedy_knapsack(x, w)
}
\arguments{
\item{x}{A DataFrame with two columns (weight(w) and value(v))}

\item{w}{Sack max weight capacity}
}
\value{
returns item with maximized value within knapsack weight limit
}
\description{
greedy_knapsack
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
